# doc/api.yaml
openapi: 3.0.3
info:
  title: Simplified login API specification
  description: |-
    This OpenAPI document describes the simplified login API.
    Copy and paste the API from the `paths` key to your OpenAPI document.
  version: "1"
...
paths:
...
  /users/{userId}/username:
    put:
      parameters:
        - name: userId
          schema:
            type: integer
            description: User identifier

          in: path
          required: true
          description: userId


      tags: ["User"]
      summary: Given the user ID, changes username of the user
      description: |-
        The logged user changes its username. It returns 0 if it has been correctly updated, or -1 if an error has occurred.

      operationId: setMyUserName

      requestBody:
        description: User name
        content:
          application/json:
            schema:
              type: string
              example: "Maria"
              readOnly: true
              minLength: 3
              maxLength: 16
              description: Username
              pattern: ^[a-zA-Z0-9]{1,20}$
        required: true

      responses:
        '200':
          description: Information related to the user
          content:
            application/json:
              schema:
                description: Information of the user; username, profile pic, followers,photos
                type: object
                properties:

                  id:
                    description: User identifier
                    type: integer
                    example: 9013213

                  name:
                    type: string
                    example: "Maria"
                    minLength: 3
                    maxLength: 16
                    pattern: ^[a-zA-Z0-9]{1,20}$
                    description: Username

                  profilePictureID:
                    description: Profile picture identifier
                    type: integer
                    example: 123912

                  banned:
                    type: array
                    description: An array with the bannedUsers.
                    minItems: 0
                    maxItems: 100
                    items:
                      type: string
                      minLength: 3
                      maxLength: 16

                  followers:
                    type: array
                    description: An array with the followers. We only store the followers, not the followed.
                    minItems: 0
                    maxItems: 100
                    items:
                      type: integer
                      description: ID of the user that follows



                  photos:
                    type: array
                    description: An array with the stream of photos.
                    minItems: 0
                    maxItems: 100
                    items:
                      type: object
                      properties:

                        imagePath:
                          description: path to the image
                          type: string
                          minLength: 3
                          maxLength: 100


                        photoID:
                          description: ID of the photo
                          type: integer

                        likes:
                          type: array
                          description: A sub-array with the likes of each photo.
                          minItems: 0
                          maxItems: 100
                          items:
                            type: integer
                            description: ID of the user that liked the photo


                        comments:
                          type: array
                          description: A sub-array with the comments of each photo.
                          minItems: 0
                          maxItems: 100
                          items:
                            type: object
                            properties:

                              user:
                                type: integer
                                description: User ID


                              comment:
                                type: string
                                minLength: 3
                                maxLength: 16
                                description: comment


                        date:
                          type: string
                          minLength: 3
                          maxLength: 16
                          format: date-time
                          description: Creation date and time


        '400':
          description: Bad Request. Error parsing data or information not valid.

        '401':
          description: Unauthorized. Missing or invalid Authentication token.



        '405':
          description: Method not allowed.
          content:
            text/plain:
              schema:
                type: string
                example: "Method Not Allowed"
                description: Error description
...
  /users/{userId}/unfollowUser/{unfollowedUserId}:
    delete:
      parameters:
        - name: userId
          schema:
            type: integer
            description: User identifier

          in: path
          required: true
          description: user logged

        - name: unfollowedUserId
          schema:
            type: integer
            description: User identifier

          in: path
          required: true
          description: user to be unfollowed




      tags: ["User"]
      summary: Given a user ID, the logged user unfollows it.
      description: |-
        The logged user unfollows the user identified by the parameter. This operation accesses to the user's followers collection and deletes it. It returns 0 if it has been correctly unfollowed, or -1 if an error has occurred (previously not followed or not found).

      operationId: unfollowUser


      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                type: integer
                example: 1
                description: Successful operation

        '401':
          description: Unauthorized. Missing or invalid Authentication token.

        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                type: integer
                example: -1
                description: Non successful operation
  /users/{userId}/unbanUser/{bannedUserId}:
    delete:
      parameters:
        - name: userId
          schema:
            type: integer
            description: User identifier

          in: path
          required: true
          description: user logged

        - name: bannedUserId
          schema:
            type: integer
            description: User identifier

          in: path
          required: true
          description: user to be unbanned




      tags: ["User"]
      summary: Given a user ID, the logged user unbans it (it deletes it from the list of banned users)
      description: |-
        The logged user unbans the user identified by the parameter. It returns 0 if it has been correctly unbanned, or -1 if an error has occurred (previously not banned or not found).

      operationId: unbanUser


      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: integer
                example: 1
                description: Successful operation

        '401':
          description: Unauthorized. Missing or invalid Authentication token.

        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                type: integer
                example: -1
                description: Non successful operation
  /users/{userId}/unlikePhoto/{photoId}:
    delete:
      parameters:
        - name: userId
          schema:
            type: integer
            description: User identifier

          in: path
          required: true
          description: user logged

        - name: photoId
          schema:
            type: integer
            description: Photo identifier

          in: path
          required: true
          description: photo to be unliked




      tags: ["Photos"]
      summary: Given a user ID and a photo Id, the logged user unlikes it
      description: |-
        The logged user deletes the logged user from the photo's list of users likes. 
        It returns 0 if it has been correctly unliked, or -1 if an error has occurred (previously not liked, userId or photoId not found).

      operationId: unlikePhoto


      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: integer
                example: 1
                description: Successful operation

        '401':
          description: Unauthorized. Missing or invalid Authentication token.

        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                type: integer
                example: -1
                description: Non successful operation
  /users/{userId}/uncommentPhoto/{photoId}:
    delete:
      parameters:
        - name: userId
          schema:
            type: integer
            description: User identifier

          in: path
          required: true
          description: user logged

        - name: photoId
          schema:
            type: integer
            description: Photo identifier

          in: path
          required: true
          description: ID of the photo to be uncommented




      tags: ["Photos"]
      summary: Given a user ID and a photo Id, the logged user uncomments it
      description: |-
        The logged user deletes the comment from the photo's list of comments. 
        It returns 0 if it has been correctly uncommented, or -1 if an error has occurred (previously not commented, userId or photoId not found).

      operationId: uncommentPhoto


      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: integer
                example: 1
                description: Successful operation

        '400':
          description: Access token is missing or invalid

        '401':
          description: Unauthorized. Missing or invalid Authentication token.

        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                type: integer
                example: -1
                description: Non successful operation
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []